requirements.txt
aiogram~=2.19
qrcode~=7.3.1
Pillow~=9.0.1
pyzbar~=0.1.8